Added UI refactorization
